typedef core classes and a bit of refactoring that introduces breaking changes
