+++ src/core/index.js
@@ -4,8 +4,6 @@
  * Module exports.
  */
 
-export * from './extrude_geometry'
-export * from './geometry_buffer'
 export * from './geometry'
 export * from './object3d'
 export * from './command'
