+++ src/geometry/index.js
@@ -4,10 +4,11 @@
  * Module exports.
  */
 
-export { TriangleGeometry } from './triangle'
-export { CylinderGeometry } from './cylinder'
-export { CapsuleGeometry } from './capsule'
-export { SphereGeometry } from './sphere'
-export { PlaneGeometry } from './plane'
-export { TorusGeometry } from './torus'
-export { BoxGeometry } from './box'
+export * from './extrude_geometry'
+export * from './triangle'
+export * from './cylinder'
+export * from './capsule'
+export * from './sphere'
+export * from './plane'
+export * from './torus'
+export * from './box'
