+++ src/math/index.js
@@ -5,6 +5,8 @@
  */
 
 export * from './quaternion'
-export * from './matrix'
-export * from './vector'
+export * from './vector1'
+export * from './vector2'
+export * from './vector3'
+export * from './vector4'
 export * from './euler'
