--- src/math/matrix.js
@@ -1,8 +1,0 @@
-'use strict'
-
-/** @TODO - implement me */
-
-export class Matrix { }
-export class Matrix2 extends Matrix {}
-export class Matrix3 extends Matrix {}
-export class Matrix4 extends Matrix {}
